+++ modules/addons/pvewhmcs/db.sql
@@ -96,8 +96,8 @@ CREATE TABLE IF NOT EXISTS `mod_pvewhmcs_plans` (
   `vlanid` varchar(10) DEFAULT NULL,
   `ipv6` varchar(10) DEFAULT 'auto',
   `balloon` varchar(10) DEFAULT '0',
+  `unpriv` int(1) unsigned DEFAULT 0,
   `ssh-keys` varchar(100) DEFAULT '',
-  `unpriv` tinyint(1) unsigned DEFAULT 0,
   PRIMARY KEY (`id`)
 );
 CREATE TABLE IF NOT EXISTS `mod_pvewhmcs_ssh_keys` (